refac: Movido e reorganizado implementações de repositories
