+++ src/main/java/com/github/edulook/look/infra/repository/StudentRepositoryAdapter.java
@@ -4,30 +4,28 @@ import com.github.edulook.look.core.model.Student;
 import com.github.edulook.look.core.repository.StudentRepository;
 import com.github.edulook.look.core.repository.student.GetStudent;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
 import java.util.Optional;
 
-@Component
+@Component("StudentRepositoryAdapter::Class")
 public class StudentRepositoryAdapter implements StudentRepository {
+    private final StudentRepository db;
+    private final StudentRepository http;
 
-    private final GetStudent getStudentClassroom;
-    private final GetStudent getGetStudentStorage;
-
-    public StudentRepositoryAdapter(@Qualifier("GetStudentClassroom::Class") GetStudent getStudentClassroom,
-                                    @Qualifier("GetStudentStorage::Class") GetStudent getGetStudentStorage) {
-        this.getStudentClassroom = getStudentClassroom;
-        this.getGetStudentStorage = getGetStudentStorage;
+    public StudentRepositoryAdapter(@Lazy @Qualifier("StudentRepositoryDB::Class") StudentRepository db,
+                                    @Lazy @Qualifier("StudentRepositoryHTTP::Class") StudentRepository http) {
+        this.db = db;
+        this.http = http;
     }
-
 
     @Override
     public Optional<Student> findStudentById(String studentId) {
-        var student = getGetStudentStorage.findStudentById(studentId);
-
+        var student = db.findStudentById(studentId);
         if(student.isEmpty())
-            return getStudentClassroom.findStudentById(studentId);
-
+            return http.findStudentById(studentId);
         return student;
     }
 }
