--- src/main/java/com/github/edulook/look/infra/repository/teacher/classroom/mapper/ClassroomTeacherAndCoreTeacherMapper.java
@@ -1,22 +1,0 @@
-package com.github.edulook.look.infra.repository.teacher.classroom.mapper;
-
-import org.springframework.stereotype.Component;
-
-import com.github.edulook.look.core.model.Teacher;
-
-import lombok.NoArgsConstructor;
-
-@Component
-@NoArgsConstructor
-public class ClassroomTeacherAndCoreTeacherMapper {
-    
-    public Teacher toModel(com.google.api.services.classroom.model.Teacher source) {
-        return Teacher
-            .builder()
-            .id(source.getUserId())
-            .name(source.getProfile().getName().getFullName())
-            .photo(source.getProfile().getPhotoUrl())
-            .email(source.getProfile().getEmailAddress())
-            .build();
-    }
-}
