+++ src/main/java/com/github/edulook/look/service/DriveService.java
@@ -17,7 +17,6 @@ import java.util.regex.Pattern;
 @Slf4j
 @Service
 public class DriveService {
-
     private final Drive drive;
 
     public DriveService(Drive drive) {
