+++ src/main/java/com/github/edulook/look/service/StudentService.java
@@ -1,22 +1,19 @@
 package com.github.edulook.look.service;
 
-import java.util.Optional;
-
+import com.github.edulook.look.core.model.Student;
+import com.github.edulook.look.core.repository.StudentRepository;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
-import com.github.edulook.look.core.model.Student;
-import com.github.edulook.look.core.repository.StudentRepository;
+import java.util.Optional;
 
-import lombok.AllArgsConstructor;
-
-/**
- * Facade student service 
- */
 @Service
-@AllArgsConstructor
 public class StudentService {
     private final StudentRepository repository;
+
+    public StudentService(StudentRepository repository) {
+        this.repository = repository;
+    }
 
     @Cacheable("getStudentProfile")
     public Optional<Student> getStudentProfile(String studentId) {
