+++ src/main/java/com/github/edulook/look/service/TeacherService.java
@@ -3,18 +3,18 @@ package com.github.edulook.look.service;
 import com.github.edulook.look.core.repository.CourseRepository;
 import com.github.edulook.look.core.repository.TeacherRepository;
 import com.github.edulook.look.endpoint.io.shared.UserAuthDTO;
-import lombok.AllArgsConstructor;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
-/**
- * Facade teacher service 
- */
 @Service
-@AllArgsConstructor
 public class TeacherService {
     private final CourseRepository courseRepository;
     private final TeacherRepository teacherRepository;
+
+    public TeacherService(CourseRepository courseRepository, TeacherRepository teacherRepository) {
+        this.courseRepository = courseRepository;
+        this.teacherRepository = teacherRepository;
+    }
 
     @Cacheable("isTeacherCourseOwner")
     public Boolean isTeacherCourseOwner(String courseId, UserAuthDTO user) {
